refactor: improve flib file structure
