--- src/bios/flib/src/disk_io/mod.rs
@@ -1,1 +1,0 @@
-pub mod disk;
