+++ src/bios/flib/src/fs/mod.rs
@@ -1,0 +1,1 @@
+pub mod partitions;
