+++ src/bios/flib/src/fs/partitions/mod.rs
@@ -1,0 +1,1 @@
+pub mod mbr;
