--- src/bios/flib/src/gdt/mod.rs
@@ -1,1 +1,0 @@
-pub mod gdt;
