+++ src/bios/flib/src/io/mod.rs
@@ -1,1 +1,3 @@
+pub mod disk;
 pub mod io;
+pub mod ps2;
