+++ src/bios/flib/src/io/ps2/mod.rs
@@ -1,18 +1,15 @@
 use core::arch::asm;
 
 pub fn send_data(data: u8) {
-
     unsafe {
         asm!(
         "out 0x60, {}",
         in(reg_byte) data
         );
     }
-
 }
 
 pub fn read_ps2() -> u8 {
-
     let data: u8;
 
     unsafe {
@@ -23,22 +20,18 @@ pub fn read_ps2() -> u8 {
     }
 
     data
-
 }
 
 pub fn send_ps2(cmd: u8) {
-
     unsafe {
         asm!(
         "out 0x64, {}",
         in(reg_byte) cmd
         );
     }
-
 }
 
 pub fn input_wait(mut loops: u16) -> Result<(), ()> {
-
     while loops > 0 {
         let status_reg: u8;
 
@@ -57,11 +50,9 @@ pub fn input_wait(mut loops: u16) -> Result<(), ()> {
     }
 
     Err(())
-
 }
 
 pub fn output_wait(mut loops: u16) -> Result<(), ()> {
-
     while loops > 0 {
         let status_reg: u8;
 
@@ -80,5 +71,4 @@ pub fn output_wait(mut loops: u16) -> Result<(), ()> {
     }
 
     Err(())
-
 }
