+++ src/bios/flib/src/lib.rs
@@ -2,17 +2,12 @@
 #![no_std]
 #[macro_use]
 
-pub mod video_io;
+pub mod video;
 pub mod bios;
-pub mod disk_io;
-pub mod gdt;
+pub mod fs;
 pub mod io;
 pub mod mem;
-pub mod part_mbr;
-pub mod ps2;
 pub mod x86;
-
-pub mod interrupts;
 
 pub use numtoa;
