+++ src/bios/flib/src/mem/mod.rs
@@ -1,2 +1,3 @@
 pub mod bmalloc;
 pub mod e820;
+pub mod gdt;
