--- src/bios/flib/src/part_mbr/mod.rs
@@ -1,1 +1,0 @@
-pub mod partition;
