--- src/bios/flib/src/ps2/mod.rs
@@ -1,1 +1,0 @@
-pub mod ps2;
