+++ src/bios/flib/src/video/vesa/video_mode.rs
