+++ src/bios/flib/src/x86/mod.rs
@@ -1,1 +1,2 @@
 pub mod flags;
+pub mod interrupts;
